requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
gunicorn==22.0.0
psycopg2-binary==2.9.9
requests==2.32.3
python-slugify==8.0.4
bleach==6.1.0
python-dotenv==1.0.1
